pyproject.toml
[tool.poetry]
name = "lms-system"
version = "0.1.0"
description = "For online study"
authors = ["Антон Федяев <fearthe54@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django = "^5.0.6"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
coverage = "^7.5.3"
drf-yasg = "^1.21.7"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.3.1"
requests = "^2.32.3"
stripe = "^9.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
